refactor(sentry): split util files, drop sync driver hook, trim comments
Addresses inline review feedback:
- Sync driver no longer needs Sentry: synchronous in-process jobs run
  inside the caller's isolation scope already, so the wrap was redundant.
- One file per export: splits sentry-workspace-context.util.ts into
  apply-workspace-sentry-context.util.ts (auth boundary) and
  apply-workspace-sentry-context-from-job-data.util.ts (queue boundary),
  with applyWorkspaceSentryFields as a small shared internal helper.
- Drops two narrative comments (in bullmq.driver.ts and instrument.ts).
