+++ packages/twenty-server/src/engine/core-modules/sentry/utils/apply-workspace-sentry-fields.util.ts
@@ -1,0 +1,26 @@
+import * as Sentry from '@sentry/node';
+
+type WorkspaceSentryFields = {
+  workspaceId: string;
+  userWorkspaceId?: string;
+};
+
+export const applyWorkspaceSentryFields = (
+  fields: WorkspaceSentryFields,
+): void => {
+  Sentry.setUser({
+    id: fields.userWorkspaceId ?? fields.workspaceId,
+  });
+
+  Sentry.setTag('twenty.workspace.id', fields.workspaceId);
+  if (fields.userWorkspaceId) {
+    Sentry.setTag('twenty.user_workspace.id', fields.userWorkspaceId);
+  }
+
+  Sentry.setContext('twenty', {
+    workspace_id: fields.workspaceId,
+    ...(fields.userWorkspaceId && {
+      user_workspace_id: fields.userWorkspaceId,
+    }),
+  });
+};
